+++ WebHook.DispatchItemStore.Client.AzureQueueStorage/AzureDispatchItemStore.cs
@@ -15,7 +15,7 @@ namespace WebHook.DispatchItemStore.Client.AzureQueueStorage
         ConcurrentDictionary<Guid, QueueMessage> inProgressMessages = new();
         public AzureDispatchItemStore()
         {
-            //TODO Config
+            //TODO make from config
             queue = new(
                 "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;",
                 "dispatch",
