+++ Webhook.Dispatcher.AzureFunction/DispatcherFunction.cs
@@ -20,13 +20,14 @@ namespace Webhook.Dispatcher.AzureFunction
         }
 
         [Function("DispatcherFunction")]
-        public async Task Run(
-            [Microsoft.Azure.Functions.Worker.QueueTrigger("dispatch", Connection = "AzureStorageQueueConnection")] string message,string PopReceipt, string id)
+        public async Task Run([QueueTrigger("dispatch", Connection = "AzureStorageQueueConnection")] string message)
         {
-            
             
             DispatchItem item = ConvertToDispatchItem(message);
 
+
+            //Functions have built in rety into host file so if failes item will be queued again after x time and then sent to DLQ after 3 attempts
+            //all set in the host.json
             //TODO catch exception types etc, probably want logic on per error code basis
             await DispatchAsync(item);
 
@@ -34,13 +35,12 @@ namespace Webhook.Dispatcher.AzureFunction
         }
         public async Task DispatchAsync(DispatchItem item)
         {
-            //TODO connection close etc
+            //TODO connection close, probably make a factory etc
             HttpClient client = new HttpClient();
             string json = JsonConvert.SerializeObject(item);
-            var stringContent = new StringContent(json, UnicodeEncoding.UTF8, "application/json");
+            StringContent stringContent = new StringContent(json, UnicodeEncoding.UTF8, "application/json");
             HttpResponseMessage result = await client.PostAsync("http://localhost:51084/Webhook", stringContent);
             result.EnsureSuccessStatusCode();
-            return;
         }
         private static DispatchItem ConvertToDispatchItem(string stringData)
         {
