+++ Webhook.Dispatcher.AzureFunction/host.json
@@ -13,9 +13,9 @@
       "messageEncoding": "base64",
       "maxPollingInterval": "00:00:02",
       "visibilityTimeout": "00:00:30",
-      "batchSize": 1,
-      "maxDequeueCount": 5,
-      "newBatchThreshold": 8
+      "batchSize": 32,
+      "maxDequeueCount": 3,
+      "newBatchThreshold": 32
     }
   }
 }
