+++ src/WebHook.Console/ProducerConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace WebHook.Producer;
+
+public partial class ProducerLoop
+{
+    public class ProducerConfig
+    {
+        public int CommitBatchSize { get; } = 50;
+    }
+}
