+++ src/WebHook.Console/ProducerLoop.cs
@@ -1,6 +1,5 @@
 ﻿using Confluent.Kafka;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 using WebHook.Contracts.Events;
 using WebHook.DispatchItemStore.Client;
 using WebHook.DispatchItemStore.Client.Redis;
@@ -9,7 +8,7 @@ using WebHook.SubscriptionStore.Client.Models;
 
 namespace WebHook.Producer;
 
-public class ProducerLoop
+public partial class ProducerLoop
 {
     private readonly ISubscriptionStore subscriptionStore;
     private readonly IDispatchItemStore dispatchItemStore;
@@ -80,40 +79,11 @@ public class ProducerLoop
         logger.LogInformation("Creating kafka consumer...");
 
         // TODO: set up proper deserializer!!
-        var builder = new ConsumerBuilder<string, IEvent>(config).SetValueDeserializer(new EventDeserilizer());
+        ConsumerBuilder<string, IEvent> builder = new ConsumerBuilder<string, IEvent>(config).SetValueDeserializer(new EventDeserilizer());
         IConsumer<string, IEvent> kafkaConsumer = builder.Build();
 
         kafkaConsumer.Subscribe(config.TopicNames);
 
         return kafkaConsumer;
-    }
-
-    public class EventDeserilizer : IDeserializer<IEvent>
-    {
-        public IEvent Deserialize(ReadOnlySpan<byte> data, bool isNull, SerializationContext context)
-        {
-            try
-            {
-                var str = System.Text.Encoding.Default.GetString(data);
-                return JsonConvert.DeserializeObject<IEvent>(str, new EventConverter());
-            }
-            catch (Exception e)
-            {
-                //TODO log fix etc
-                return null;
-            }
-          
-        }
-    }
-    public class EventLogConsumerConfig : ConsumerConfig
-    {
-        public IReadOnlyList<string> TopicNames { get; }
-
-        public EventLogConsumerConfig(IReadOnlyList<string> topicNames) => TopicNames = topicNames;
-    }
-
-    public class ProducerConfig
-    {
-        public int CommitBatchSize { get; } = 50;
     }
 }
