+++ webhook-prototype.sln
@@ -39,7 +39,17 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MockWebHookEndpoint", "Mock
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MockKafkaEventProducer", "MockKafkaEventProducer\MockKafkaEventProducer.csproj", "{AB025607-B20E-4699-A4B9-262243DD5F20}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webhook.Dispatcher.AzureFunction", "Webhook.Dispatcher.AzureFunction\Webhook.Dispatcher.AzureFunction.csproj", "{17EDE83C-CD2B-4822-863F-579AF06DF295}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Webhook.Dispatcher.AzureFunction", "Webhook.Dispatcher.AzureFunction\Webhook.Dispatcher.AzureFunction.csproj", "{17EDE83C-CD2B-4822-863F-579AF06DF295}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Dispatchers", "Dispatchers", "{55D4DC07-49EB-4F1B-8880-61389B29B609}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{4782EABE-5B4B-4B30-AC62-28A315A81B30}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Producers", "Producers", "{3B3C26A3-D824-409F-96BA-C587F3E02B7C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{8D207B20-885E-4BFC-8FC7-30489A09C3B2}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{A3F9A70F-0000-4DB9-A180-1AB2C61F9FEA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -100,13 +110,13 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{76931A20-1E7A-4646-A4CF-11BFFF91B87B} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
-		{09C1F83D-9431-4518-96BC-DC23E105C870} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
-		{0B1D2074-308D-42BD-B255-848DB8AB59D8} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
-		{28EC3118-ABAB-4766-8D12-FE0468BFA94C} = {489785E4-03C5-4100-816C-7CA751D970C6}
+		{76931A20-1E7A-4646-A4CF-11BFFF91B87B} = {3B3C26A3-D824-409F-96BA-C587F3E02B7C}
+		{09C1F83D-9431-4518-96BC-DC23E105C870} = {4782EABE-5B4B-4B30-AC62-28A315A81B30}
+		{0B1D2074-308D-42BD-B255-848DB8AB59D8} = {55D4DC07-49EB-4F1B-8880-61389B29B609}
+		{28EC3118-ABAB-4766-8D12-FE0468BFA94C} = {A3F9A70F-0000-4DB9-A180-1AB2C61F9FEA}
 		{9BA6A5D7-7FC7-46A7-9A25-62835AC9F92A} = {489785E4-03C5-4100-816C-7CA751D970C6}
 		{489785E4-03C5-4100-816C-7CA751D970C6} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
-		{6EC32576-C3A1-4329-A523-CF097E71E9E3} = {80246D97-5A80-4AFE-AF2E-54320DA6BC7B}
+		{6EC32576-C3A1-4329-A523-CF097E71E9E3} = {8D207B20-885E-4BFC-8FC7-30489A09C3B2}
 		{80246D97-5A80-4AFE-AF2E-54320DA6BC7B} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
 		{6795D3F9-B8B7-4D2B-9AAE-65E302ABD7AD} = {80246D97-5A80-4AFE-AF2E-54320DA6BC7B}
 		{8BF3AAB8-1551-4456-9AB3-291AB0211F39} = {489785E4-03C5-4100-816C-7CA751D970C6}
@@ -114,7 +124,12 @@ Global
 		{F00FC014-5E87-450B-9204-86829F8D5B7A} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
 		{3A2DC8A7-BDD7-454F-85F6-1F17F4B4A724} = {F00FC014-5E87-450B-9204-86829F8D5B7A}
 		{AB025607-B20E-4699-A4B9-262243DD5F20} = {F00FC014-5E87-450B-9204-86829F8D5B7A}
-		{17EDE83C-CD2B-4822-863F-579AF06DF295} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
+		{17EDE83C-CD2B-4822-863F-579AF06DF295} = {55D4DC07-49EB-4F1B-8880-61389B29B609}
+		{55D4DC07-49EB-4F1B-8880-61389B29B609} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
+		{4782EABE-5B4B-4B30-AC62-28A315A81B30} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
+		{3B3C26A3-D824-409F-96BA-C587F3E02B7C} = {4731DDE3-47DF-4EC8-BA71-34A012C67060}
+		{8D207B20-885E-4BFC-8FC7-30489A09C3B2} = {80246D97-5A80-4AFE-AF2E-54320DA6BC7B}
+		{A3F9A70F-0000-4DB9-A180-1AB2C61F9FEA} = {489785E4-03C5-4100-816C-7CA751D970C6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {318DA05C-C93C-41E9-BC17-9D053EFB1CC3}
